Restructured LandingPage components.  Working point before I mess with Modals
