+++ client/src/components/Billing/BillingForm/index.js
@@ -1,0 +1,2 @@
+import BillingForm from "./BillingForm.js";
+export default BillingForm;
