+++ client/src/components/Billing/BillingView/index.js
@@ -1,0 +1,2 @@
+import BillingView from "./BillingView.js";
+export default BillingView;
