--- client/src/components/Billing/Form/index.js
@@ -1,2 +1,0 @@
-import Form from "./Form.js";
-export default Form;
