--- client/src/components/Billing/View/index.js
@@ -1,2 +1,0 @@
-import View from "./View.js";
-export default View;
