+++ client/src/components/Callback/callback.js
@@ -1,13 +1,14 @@
 import React, { useEffect } from "react";
 
 //Loading SVG that gets displayed
-import Progress from "components/Gen/Progress/ProgressCircle";
+import Progress from "components/Misc/Progress/ProgressCircle";
 
 //Auth
 import { setAccessToken, setIdToken, getUserProfile } from "Auth/Auth";
 
 function Callback(props) {
   useEffect(() => {
+    // componentDidMount
     setAccessToken();
     setIdToken();
     getUserProfile(() => {
