+++ client/src/components/LandingPage/LandingPageView/index.js
@@ -1,0 +1,2 @@
+import LandingPageView from "./LandingPageView.js";
+export default LandingPageView;
