+++ client/src/components/LandingPage/Pricing/index.js
@@ -1,0 +1,2 @@
+import Pricing from "./Pricing.js";
+export default Pricing;
