+++ client/src/components/LandingPage/Team/index.js
@@ -1,0 +1,2 @@
+import Team from "./Team.js";
+export default Team;
