--- client/src/components/LandingPage/index.js
@@ -1,2 +1,0 @@
-import LandingPageView from "./LandingPageView";
-export default LandingPageView;
