+++ client/src/components/TeamMembers/TeamMemberPageContainer/TeamMemberPageView.js
@@ -3,7 +3,7 @@ import React from "react";
 //Components
 import TeamMemberPage from "./TeamMemberPage";
 import Snackbar from "../../Snackbar/Snackbar";
-import ProgressCircle from "../../Progress/ProgressCircle";
+import ProgressCircle from "components/Misc/Progress/ProgressCircle.js";
 
 //Redux
 import { connect } from "react-redux";
