+++ client/src/routes.js
@@ -1,28 +1,28 @@
-import React from 'react';
+import React from "react";
 
 //Routing
-import { Route, Router } from 'react-router-dom';
+import { Route, Router } from "react-router-dom";
 
 //Components
-import App from './App';
-import Pricing from './components/LandingPage/Pricing';
-import Team from './components/LandingPage/Team';
-import DashboardView from './components/Dashboard/DashboardView';
+import App from "./App";
+import Pricing from "components/LandingPage/Pricing/Pricing.js";
+import Team from "./components/LandingPage/Team/";
+import DashboardView from "./components/Dashboard/DashboardView";
 
 //Callback
-import Callback from './components/Callback/callback';
+import Callback from "./components/Callback/Callback.js";
 
 //Auth
-import Authenticate from './components/authenticate/authenticate';
+import Authenticate from "components/Misc/authenticate/authenticate.js";
 
 //History
-import history from './history';
+import history from "./history";
 
-import { Provider } from 'react-redux';
-import { createStore, applyMiddleware, compose } from 'redux';
-import thunk from 'redux-thunk';
+import { Provider } from "react-redux";
+import { createStore, applyMiddleware, compose } from "redux";
+import thunk from "redux-thunk";
 
-import rootReducer from './store/reducers';
+import rootReducer from "./store/reducers";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
